Refactor async & redux
