+++ src/modules/activities/models/models.ts
@@ -1,0 +1,12 @@
+export type UserActivity = {
+  date: Date;
+  title: string;
+  id: string;
+  score: number;
+};
+
+export interface ActivitiesState {
+  isLoading: boolean;
+  userActivities?: UserActivity[];
+  error?: string;
+}
