+++ src/modules/app/index.ts
@@ -1,3 +1,4 @@
 export * from './components';
 export * from './redux';
 export * from './hooks';
+export * from './models';
