+++ src/modules/app/redux/reducer.ts
@@ -1,18 +1,4 @@
-import { AppActionTypes, AppActions } from 'modules/app';
-export interface Theme {
-  color: string;
-  showNav: boolean;
-  shapeClass: string;
-}
-
-export interface AppVariables {
-  theme: Theme;
-}
-
-export type AppState = AppVariables & {
-  isLoading: boolean;
-  error?: string;
-};
+import { AppActionTypes, AppActions, AppState } from 'modules/app';
 
 const INITIAL_STATE: AppState = {
   theme: {
