+++ src/modules/contacts/index.ts
@@ -2,3 +2,4 @@ export * from './services';
 export * from './components';
 export * from './redux';
 export * from './hooks';
+export * from './models';
