+++ src/modules/contacts/models/models.ts
@@ -1,0 +1,12 @@
+export type UserContact = {
+  date: Date;
+  name: string;
+  id: string;
+  phoneNumber: string;
+};
+
+export interface ContactsState {
+  isLoading: boolean;
+  userContacts?: UserContact[];
+  error?: string;
+}
