+++ src/modules/score/index.ts
@@ -2,3 +2,4 @@ export * from './redux';
 export * from './components';
 export * from './services';
 export * from './hooks';
+export * from './models';
