+++ src/modules/settings/index.ts
@@ -3,3 +3,4 @@ export * from './services';
 export * from './redux';
 export * from './constants';
 export * from './hooks';
+export * from './models';
