+++ src/modules/settings/models/models.ts
@@ -1,0 +1,7 @@
+export type UserSettings = { [key: string]: boolean };
+
+export interface SettingsState {
+  isLoading: boolean;
+  userSettings?: UserSettings;
+  error?: string;
+}
