--- src/modules/user/models/User.ts
@@ -1,4 +1,0 @@
-export interface UserAuthData {
-  email: string;
-  password: string;
-}
