+++ src/modules/user/models/models.ts
@@ -1,0 +1,10 @@
+export interface UserAuthData {
+  email: string;
+  password: string;
+}
+
+export interface UserState {
+  isLoading: boolean;
+  userData?: firebase.UserInfo;
+  error?: string;
+}
